Refine audio imports for merge compatibility
