+++ OcchioOnniveggente/src/storage/__init__.py
@@ -1,0 +1,5 @@
+from .metadata import MetadataStore
+from .vector import VectorStore
+
+__all__ = ["MetadataStore", "VectorStore"]
+
